base: add Area Detector core modules.
Include only core modules from Area Detector, which are required to
build any other submodule.

Git has been included in the build tools, since it allows to easily
download the Area Detector modules in a single command in a parallel
fashion. This makes easier to define the compatible versions as well,
since it is defined at the supermodule level. For this configuration,
the latest release is being used.

Since area detector configuration is quite unique when compared to other
EPICS modules, a new configuration script has been included for setting
it up. This basically configures the top-level supermodule, which
recursively compiles all modules listed in the `RELEASE.local` file.
This configuration mostly follows the ones provided, and thus uses `sed`
to edit the paths and the options instead of having the whole
`configure` files written at once.

Note that the IOCs inside the area detector modules have not been built,
since this docker image serves for the purpose of being the support
building the IOCs themselves in other images.

Several libs from ADSupport has been already configured to be built (as
it is the default value). The exception here is MagicGraphicK, because
it is only required by ADFileMagick but requires some system libraries
which aren't really needed to be handled for now.