base: compile ADSimDetector in Area Detector module.
ADSimDetector requires busy and sscan, which have been now included in
the build.

Note that we are not building the IOC in this container, since it is the
support for having the IOCs built. This includes the ADSimDetector.
Thus, another Docker container is recommended to be used to have the
simulator IOC build using the library compiled in this image. This
should be straight-forward, as it should simply start a IOC shell and
define the start-up commands (st.cmd).